moved everything related to change-types, approvers and bundle files
this functionality is kept free from details bound to roles and avoid being concrete about where how a changetype gets applied to a change.
